+++ tests/test_types_release_cookies_full.c
@@ -6,7 +6,7 @@ void test_impl()
 {
   struct HSCookies *cookies = hs_types_new_cookies(1);
 
-  for (size_t index = 0; index < cookies->count; index++)
+  for (size_t index = 0; index < 10; index++)
   {
     struct HSCookie *cookie = hs_types_new_cookie();
     cookie->name = stringfn_new_empty_string();