+++ tests/test_types_release_key_value_array_full.c
@@ -6,10 +6,9 @@ void test_impl()
 {
   struct HSKeyValueArray *key_value_array = hs_types_new_key_value_array(50);
 
-  key_value_array->count = 5;
-  for (size_t index = 0; index < key_value_array->count; index++)
+  for (size_t index = 0; index < 10; index++)
   {
-    key_value_array->pairs[index] = hs_types_new_key_value(stringfn_new_empty_string(), stringfn_new_empty_string());
+    hs_types_key_value_array_add(key_value_array, stringfn_new_empty_string(), stringfn_new_empty_string());
   }
 
   hs_types_release_key_value_array(key_value_array);