+++ tests/test_types_release_post_response_callbacks.c
@@ -1,0 +1,31 @@
+#include "stringfn.h"
+#include "test.h"
+
+
+void _test_release(struct HSPostResponseCallback *callback)
+{
+  hs_io_free(callback->context);
+}
+
+
+void test_impl()
+{
+  struct HSPostResponseCallbacks *callbacks = hs_types_new_post_response_callbacks(1);
+
+  for (size_t index = 0; index < 10; index++)
+  {
+    struct HSPostResponseCallback *callback = hs_types_new_post_response_callback();
+    callback->context = stringfn_new_empty_string();
+    callback->release = _test_release;
+    hs_types_post_response_callbacks_add(callbacks, callback);
+  }
+
+  hs_types_release_post_response_callbacks(callbacks);
+}
+
+
+int main()
+{
+  test_run(test_impl);
+}
+