+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Refactor router flow invocation to enable non response routes to modify final HTTP response.
 * New simplified api for adding headers and cookies
 * New powered-by route
+* Support multiple post response callbacks
 
 ### v0.2.3 (2021-08-24)